requirements.txt
fastapi #~=0.116.1
streamlit #~=1.47.1
langgraph #~=0.6.3
langchain-openai #~=1.1.2
httpx #~=0.28.1
python-dotenv #~=1.1.1
langchain-core #>=1.1.3
unidecode #~=1.4.0
pydantic #~=2.11.7
starlette #~=0.47.2
certifi #~=2025.8.3
requests #~=2.32.3
uvicorn
langchain
langchain-tavily
